refactor: rename to DataTable
