+++ packages/components/data-table/src/index.js
@@ -1,3 +1,3 @@
-export { default } from './simple-table';
+export { default } from './data-table';
 export { useRowSelection, useSorting } from '../../../hooks';
 export { default as version } from './version';
